feat: actionscell 분리
