+++ src/features/posts/ui/table/PostsTableRow.tsx
@@ -1,14 +1,11 @@
-import { Button, TableCell, TableRow } from "../../../../shared/ui"
+import { TableCell, TableRow } from "../../../../shared/ui"
 import { Highlight } from "../../../../shared/ui/Highlight.tsx"
-import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { ThumbsDown, ThumbsUp } from "lucide-react"
 import React from "react"
 import { Post } from "../../../../entities/posts/model/types.ts"
-import { usePostsStore } from "../../model/postsStore.ts"
-import { usePostDialogStore } from "../../model/postDialogStore.ts"
 import { useQueryParams } from "../../model/useQueryParams.ts"
-import { useCommentsStore } from "../../../comments/model/commentStore.ts"
-import { useDeletePostMutation } from "../../api/queries.ts"
 import { UserModalCell } from "./UserModalCell.tsx"
+import { ActionsCell } from "./ActionsCell.tsx"
 
 interface PostsTableRowProps {
   post: Post
@@ -16,41 +13,6 @@ interface PostsTableRowProps {
 
 export const PostsTableRow: React.FC<PostsTableRowProps> = ({ post }) => {
   const { searchQuery, selectedTag, setSelectedTag, updateURL } = useQueryParams()
-  const posts = usePostsStore((state) => state.posts)
-  const { setPosts, setSelectedPost } = usePostsStore()
-
-  const comments = useCommentsStore((state) => state.comments)
-  const { setComments } = useCommentsStore()
-
-  const { setShowPostEditDialog, setShowPostDetailDialog } = usePostDialogStore()
-
-  const { mutate: deletePostMutation } = useDeletePostMutation()
-
-  const fetchComments = async (postId: number) => {
-    if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
-    try {
-      const response = await fetch(`/api/comments/post/${postId}`)
-      const data = await response.json()
-      setComments((prev) => ({ ...prev, [postId]: data.comments }))
-    } catch (error) {
-      console.error("댓글 가져오기 오류:", error)
-    }
-  }
-
-  // 게시물 상세 보기
-  const openPostDetail = (post: Post) => {
-    setSelectedPost(post)
-    fetchComments(post.id)
-    setShowPostDetailDialog(true)
-  }
-
-  const deletePost = async (id: number) => {
-    deletePostMutation(id, {
-      onSuccess: () => {
-        setPosts(posts.filter((post) => post.id !== id))
-      },
-    })
-  }
 
   return (
     <TableRow key={post.id}>
@@ -90,26 +52,7 @@ export const PostsTableRow: React.FC<PostsTableRowProps> = ({ post }) => {
           <span>{post.reactions?.dislikes || 0}</span>
         </div>
       </TableCell>
-      <TableCell>
-        <div className="flex items-center gap-2">
-          <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
-            <MessageSquare className="w-4 h-4" />
-          </Button>
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => {
-              setSelectedPost(post)
-              setShowPostEditDialog(true)
-            }}
-          >
-            <Edit2 className="w-4 h-4" />
-          </Button>
-          <Button variant="ghost" size="sm" onClick={() => deletePost(post.id)}>
-            <Trash2 className="w-4 h-4" />
-          </Button>
-        </div>
-      </TableCell>
+      <ActionsCell post={post} />
     </TableRow>
   )
 }
